[remSensLocal.ino]
Changed makeString() to makeJson().
Have to reduce lstTemp and lstTime size to allow encoded to be created correctly

encodedJson created
recives the json string to be URL encoded and returns an encoded String
near 600 bytes used to perform this action.

Need to improve to use only c strings, sprintf and strcat.